kafka: add seq_stm to validate seq numbers within a session
Idempotent producer feature is based on using a pair of pid & epoch
to identify a session and the sequential (seq) numbers to identify
a message within a session. seq_stm keeps the last used seq number
per session and enforces that the seq numbers within a session form
a increasing sequence without the gaps.