r/consensus: make a snapshot before eviction
We need to take snapshots before eviction to avoid information loss.

Example:

Let a state machine sum a stream of numbers by key so the following sequence

    (k1,1), (k2,3), (k1,2), (k3,7)

resolves to

    { "k1": 3, "k2": 3, "k3": 7 }

If we evict the log up to 3rd offset then the sequence turns into

    _, _, _, |, (k3,7)

Where "|" marks the eviction moment. When we replay the sequence it resolves
to

    { "k3": 7 }

which differs from the original state. Saving a snapshot before the eviction
solves this problem:

    _, _, _, |, (k3,7)
    { "k1": 3, "k2": 3 }

When we replay the log starting from the last snapshot it leads to the original
state:

    { "k1": 3, "k2": 3, "k3": 7 }