r/consensus: make a snapshot before compaction
We need to take snapshots before compaction to avoid information loss.

Example:

Let a state machine sum a stream of numbers by key so the following sequence

    (k1,1), (k2,3), (k1,2), (k3,7)

resolves to

    { "k1": 3, "k2": 3, "k3": 7 }

If we compact the log up to 3rd offset then the sequence turns into

    _, (k2,3), (k1,2), |, (k3,7)

Where "|" marks the compaction moment. When we replay the sequence it resolves
to

    { "k1": 2, "k2": 3, "k3": 7 }

which differs from the original state. Saving a snapshot before the compaction
solves this problem:

    _, (k2,3), (k1,2), |, (k3,7)
    { "k1": 3, "k2": 3 }

When we replay the log starting from the last snapshot it leads to the original
state:

    { "k1": 3, "k2": 3, "k3": 7 }

Compaction happens in phases. During the first phase we compact the segments
and then we do cross-segment compaction of already compacted segments. Since we
makes snapshot before compacting each segment compaction then when things come
to cross-segment compaction there is no need to take a snapshot because the
snapshot of the last compacted segment already covers the affected offset range.